+++ netpyne/simFuncs.py
@@ -367,6 +367,23 @@ def _loadFile (filename):
     import sim
     import os
 
+    def _byteify(data, ignore_dicts = False):
+        # if this is a unicode string, return its string representation
+        if isinstance(data, unicode):
+            return data.encode('utf-8')
+        # if this is a list of values, return list of byteified values
+        if isinstance(data, list):
+            return [ _byteify(item, ignore_dicts=True) for item in data ]
+        # if this is a dictionary, return dictionary of byteified keys and values
+        # but only if we haven't already byteified it
+        if isinstance(data, dict) and not ignore_dicts:
+            return OrderedDict({
+                _byteify(key, ignore_dicts=True): _byteify(value, ignore_dicts=True)
+                for key, value in data.iteritems()
+            })
+        # if it's anything else, return it in its original form
+        return data
+
     if hasattr(sim, 'cfg') and sim.cfg.timing: sim.timing('start', 'loadFileTime')
     ext = os.path.basename(filename).split('.')[1]
 
@@ -390,7 +407,7 @@ def _loadFile (filename):
         import json
         print('Loading file %s ... ' % (filename))
         with open(filename, 'r') as fileObj:
-            data = json.load(fileObj, object_pairs_hook=OrderedDict)
+            data = json.load(fileObj, object_hook=_byteify)
 
     # load mat file
     elif ext == 'mat':